refactor: rename tilte store file to document
