+++ src/lib/stores/document.ts
