+++ src/lib/stores/index.ts
@@ -2,4 +2,4 @@ export * from "./accounts";
 export * from "./deals";
 export * from "./inventory";
 export * from "./selected";
-export * from "./title";
+export * from "./document";
